Split crop into its own develop sidebar panel
Move crop controls out of EditPanel into a dedicated CropPanel with its
own rail button. Edit panel now only shows basic, curve, mixer, and
effects adjustments.
