+++ components/develop/EditPanel.tsx
@@ -6,6 +6,8 @@ import type { DevelopPluginId, MixerColor } from "@/lib/develop/types";
 import { useDevelopStore } from "@/stores/develop-store";
 import { SliderRow } from "@/components/develop/SliderRow";
 
+const EDIT_PLUGINS = DEVELOP_PLUGINS.filter((plugin) => plugin.id !== "crop");
+
 const MIXER_LABELS: Record<MixerColor, string> = {
   red: "Red",
   orange: "Orange",
@@ -44,7 +46,7 @@ export function EditPanel() {
       </div>
 
       <div className="flex-1 overflow-auto">
-        {DEVELOP_PLUGINS.map((plugin) => (
+        {EDIT_PLUGINS.map((plugin) => (
           <PluginSection key={plugin.id} id={plugin.id} label={plugin.label} />
         ))}
       </div>
@@ -67,7 +69,6 @@ function PluginSection({
         {label}
       </summary>
       <div className="px-3 pb-3">
-        {id === "crop" ? <CropControls /> : null}
         {id === "basic" ? <BasicControls /> : null}
         {id === "curve" ? <CurveControls /> : null}
         {id === "mixer" ? <MixerControls /> : null}
@@ -81,34 +82,6 @@ function PluginSection({
         </button>
       </div>
     </details>
-  );
-}
-
-function CropControls() {
-  const crop = useDevelopStore((state) => state.settings.crop);
-  const updatePlugin = useDevelopStore((state) => state.updatePlugin);
-
-  return (
-    <div>
-      <label className="mb-2 flex items-center gap-2 text-xs text-lr-text-muted">
-        <input
-          type="checkbox"
-          checked={crop.enabled}
-          onChange={(event) =>
-            updatePlugin("crop", { enabled: event.target.checked })
-          }
-        />
-        Enable crop
-      </label>
-      <SliderRow label="Left" value={crop.x} min={0} max={0.95} step={0.01} onChange={(x) => updatePlugin("crop", { x })} />
-      <SliderRow label="Top" value={crop.y} min={0} max={0.95} step={0.01} onChange={(y) => updatePlugin("crop", { y })} />
-      <SliderRow label="Width" value={crop.width} min={0.05} max={1} step={0.01} onChange={(width) => updatePlugin("crop", { width })} />
-      <SliderRow label="Height" value={crop.height} min={0.05} max={1} step={0.01} onChange={(height) => updatePlugin("crop", { height })} />
-      <SliderRow label="Straighten" value={crop.angle} min={-45} max={45} step={0.1} suffix="deg" onChange={(angle) => updatePlugin("crop", { angle })} />
-      <SliderRow label="Perspective X" value={crop.perspectiveX} min={-100} max={100} onChange={(perspectiveX) => updatePlugin("crop", { perspectiveX })} />
-      <SliderRow label="Perspective Y" value={crop.perspectiveY} min={-100} max={100} onChange={(perspectiveY) => updatePlugin("crop", { perspectiveY })} />
-      <SliderRow label="Distortion" value={crop.distortion} min={-100} max={100} onChange={(distortion) => updatePlugin("crop", { distortion })} />
-    </div>
   );
 }
